rename test files & move tests to different files
